fix: refactor tests for the format validators
