+++ tests/integration/components/cf-field-test.js
@@ -3,13 +3,15 @@ import { setupRenderingTest } from "ember-qunit";
 import { render, fillIn } from "@ember/test-helpers";
 import hbs from "htmlbars-inline-precompile";
 import Document from "ember-caluma/lib/document";
-import ValidatorServiceStub from "dummy/tests/helpers/validator-service-stub";
+import { settled } from "@ember/test-helpers";
+import setupMirage from "ember-cli-mirage/test-support/setup-mirage";
 
 module("Integration | Component | cf-field", function(hooks) {
   setupRenderingTest(hooks);
+  setupMirage(hooks);
 
   hooks.beforeEach(function() {
-    this.owner.register("service:validator", ValidatorServiceStub);
+    this.server.create("format-validator", { slug: "email", regex: "/@/" });
 
     const form = {
       __typename: "Form",
@@ -21,6 +23,16 @@ module("Integration | Component | cf-field", function(hooks) {
           isRequired: "true",
           isHidden: "true",
           textMaxLength: 2,
+          __typename: "TextQuestion"
+        },
+        {
+          slug: "question-2",
+          label: "Test2",
+          isRequired: "true",
+          isHidden: "true",
+          meta: {
+            formatValidators: ["email"]
+          },
           __typename: "TextQuestion"
         }
       ]
@@ -45,6 +57,7 @@ module("Integration | Component | cf-field", function(hooks) {
     });
 
     this.set("field", document.fields[0]);
+    this.set("emailField", document.fields[1]);
   });
 
   test("it allows deleting existing input", async function(assert) {
@@ -96,5 +109,29 @@ module("Integration | Component | cf-field", function(hooks) {
     await render(hbs`{{cf-field field=field}}`);
 
     assert.dom("uk-text-bold").doesNotExist();
+  });
+
+  test("it shows error message", async function(assert) {
+    assert.expect(1);
+    let service = this.owner.lookup("service:validator");
+    await settled();
+    let error = service.validators.find(i => i.slug === "email").errorMsg;
+
+    await render(hbs`{{cf-field field=emailField}}`);
+
+    await fillIn("input", "Test");
+
+    assert.dom("span.validation-errors").hasText(error);
+  });
+
+  test("it saves the valid email address", async function(assert) {
+    assert.expect(1);
+    await settled();
+
+    await render(hbs`{{cf-field field=emailField}}`);
+
+    await fillIn("input", "test@test.com");
+
+    assert.dom("span.validation-errors").doesNotExist();
   });
 });
