+++ tests/integration/components/cfb-form-editor/question/validation-test.js
@@ -1,10 +1,9 @@
 import { module, test } from "qunit";
 import { setupRenderingTest } from "ember-qunit";
-import { render, fillIn } from "@ember/test-helpers";
-import Document from "ember-caluma/lib/document";
+import { render } from "@ember/test-helpers";
 import hbs from "htmlbars-inline-precompile";
-import { setupMirage } from "ember-cli-mirage/test-support";
-import { settled } from "@ember/test-helpers";
+import setupMirage from "ember-cli-mirage/test-support/setup-mirage";
+import { click } from "@ember/test-helpers";
 
 module(
   "Integration | Component | cfb-form-editor/question/validation",
@@ -13,70 +12,18 @@ module(
     setupMirage(hooks);
 
     hooks.beforeEach(function() {
-      this.server.create("format-validator", { slug: "email" });
-
-      const form = {
-        __typename: "Form",
-        slug: "some-form",
-        questions: [
-          {
-            slug: "question-1",
-            label: "Test",
-            isRequired: "true",
-            isHidden: "false",
-            meta: {
-              formatValidators: ["email"]
-            },
-            __typename: "TextQuestion"
-          }
-        ]
-      };
-
-      const document = Document.create(this.owner.ownerInjection(), {
-        raw: {
-          id: window.btoa("Document:1"),
-          __typename: "Document",
-          answers: [
-            {
-              stringValue: "Test",
-              question: {
-                slug: "question-1"
-              },
-              __typename: "StringAnswer"
-            }
-          ],
-          rootForm: form,
-          forms: [form]
-        }
-      });
-
-      this.set("field", document.fields[0]);
+      this.server.createList("format-validator", 5);
     });
 
-    test("it shows error message", async function(assert) {
-      assert.expect(1);
-      let service = this.owner.lookup("service:validator");
-      await settled();
-      let error = service.validators.find(i => i.slug === "email").errorMsg;
+    test("it renders all available validators", async function(assert) {
+      assert.expect(2);
 
-      await render(hbs`{{cf-field field=field}}`);
+      await render(hbs`{{cfb-form-editor/question/validation}}`);
 
-      await fillIn("input", "Test");
+      await click(".ember-power-select-trigger-multiple-input");
 
-      assert.dom("span.validation-errors").hasText(error);
-    });
-
-    test("it saves the valid email address", async function(assert) {
-      assert.expect(1);
-      let service = this.owner.lookup("service:validator");
-      await settled();
-      let error = service.validators.find(i => i.slug === "email").errorMsg;
-
-      await render(hbs`{{cf-field field=field}}`);
-
-      await fillIn("input", "test@test.com");
-
-      assert.dom("span.validation-errors").hasText(error);
+      assert.dom(".ember-power-select-option").exists({ count: 5 });
+      assert.dom(".ember-power-select-option").hasText("Validator #1");
     });
   }
 );
